Minor useability tweak
Modify pattern svg to avoid long vertical spears in the design that
might get caught on things.

This is not in the KiCad, so it would have to be re-done if any
changes are made there (although it might be possible to include
it in the KiCad artwork if it becomes a headache).